+++ packages/babel-plugin-transform-jsx-stylesheet/src/index.ts
@@ -171,16 +171,7 @@ export default function ({ types: t, template }, opts = {}) {
           }
 
           if (isCssModule) {
-            classNameAttribute.name.name = 'style';
-            // `className` is already transformed to `style`
-            // `hasStyleAttribute` should equal `true`
-            // and `styleAttribute` should equal `classNameAttribute`
-            if (retainClassName) {
-              hasStyleAttribute = true;
-              styleAttribute = classNameAttribute;
-              hasClassName = false;
-              classNameAttribute = t.jSXAttribute(t.jSXIdentifier('className'), null);
-            }
+            attributes.splice(attributes.indexOf(classNameAttribute), 1);
             file.set('injectGetStyle', false);
           }