+++ packages/plugin-ssr/src/ssr/addBuiltInRenderComponentToHTML.ts
@@ -3,18 +3,28 @@ import addPageHTMLAssign from './addPageHTMLAssign';
 
 export default function addBuiltInRenderComponentToHTML({ updateDataInClient }: IFormattedLoaderQuery) {
   return `
-  async function renderComponentToHTML(Component, ctx, initialData, htmlTemplate) {
+  async function renderComponentToHTML(Component, ctx, options ) {
+    const { initialData, htmlTemplate, initialProps } = options;
     const $ = new Generator(htmlTemplate || '__RAX_APP_SERVER_HTML_TEMPLATE__');
 
-    const pageInitialProps = await getInitialProps(Component, ctx);
+    const pageInitialProps = initialProps || await getInitialProps(Component, ctx);
     const data = {
       __SSR_ENABLED__: true,
       initialData,
       pageInitialProps
     };
-    // Assign pageHTML
-    ${addPageHTMLAssign()}
-    $.root.innerHTML = pageHTML;
+
+    let initialHtml;
+
+    if (options.__pageHTML != null) {
+      initialHtml = options.__pageHTML;
+    } else {
+      // Assign pageHTML
+      ${addPageHTMLAssign()}
+      initialHtml = pageHTML;
+    }
+
+    $.root.innerHTML = initialHtml;
     ${updateDataInClient ? '' : '$.insertScript(\'<script data-from="server">window.__INITIAL_DATA__=\' + stripXSS(JSON.stringify(data)) + \'</script>\');'}
     return $.html();
   };